requirements_v0.3.6_AI_20260222_2322.txt
affine==2.4.0
aggdraw==1.4.1
aiobotocore==3.1.2
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
alabaster==1.0.0
anamnesis==1.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asttokens==3.0.1
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
boto3==1.42.42
botocore==1.42.42
Cartopy==0.23.0
certifi==2026.1.4
cffi==2.0.0
cftime==1.6.5
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.2
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
Cython==3.2.4
dask==2026.1.2
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
docutils==0.22.4
donfig==0.8.1.post1
emoji==2.15.0
executing==2.2.1
filetype==1.2.0
fiona==1.10.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
geodatasets==2026.1.0
geopandas==1.1.2
glmtools==0.3.0
-e git+ssh://git@github.com/deliaspanigo/MAIE_tesis_github.git@b1f6f5869f01e8350e85f07a90078483699e2963#egg=goes_processor
-e git+ssh://git@github.com/deliaspanigo/goes19-sat-processor.git@3a768a98c61b0be62fc6c0b8c8bc978782bd78f5#egg=goes19_processor
google-crc32c==1.8.0
h11==0.16.0
h5netcdf==1.8.1
h5py==3.15.1
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imagesize==1.4.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
langdetect==1.0.9
llvmlite==0.46.0
locket==1.0.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.8.4
matplotlib-inline==0.2.1
multidict==6.7.1
nest-asyncio==1.6.0
netCDF4==1.7.4
nltk==3.9.2
numba==0.64.0
numcodecs==0.16.5
numpy==2.4.2
numpydoc==1.10.0
olefile==0.47
packaging==26.0
pandas==3.0.0
parso==0.8.6
partd==1.4.2
pathlib_abc==0.5.2
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.6.0
pooch==1.9.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycoast==1.8.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pykdtree==1.4.3
pyogrio==0.12.1
pyorbital==1.12.0
pyparsing==3.3.2
pypdf==6.7.2
pypdfium2==5.5.0
pyproj==3.7.2
pyresample==1.35.0
pyshp==3.0.3
pyspectral==0.14.1
python-dateutil==2.9.0.post0
python-geotiepoints==1.9.0
python-iso639==2026.1.31
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
rasterio==1.5.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rioxarray==0.21.0
roman-numerals==4.1.0
s3fs==2026.2.0
s3transfer==0.16.0
satpy==0.59.0
scipy==1.17.0
setuptools==82.0.0
shapely==2.1.2
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.8.3
Sphinx==9.1.0
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
trollimage==1.28.0
trollsift==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
universal_pathlib==0.3.9
unstructured==0.20.8
unstructured-client==0.42.10
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
wheel==0.46.3
wrapt==1.17.3
xarray==2025.11.0
yarl==1.22.0
zarr==3.1.5
